@@ -6,7 +6,7 @@ description: Demonstrates how to use the firebase_app_indexing plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <3.5.0"
 
 dependencies:
   flutter:
