@@ -1,10 +1,10 @@
 name: firebase_app_indexing
-description: Firebase app indexing plugin, 
+description: Firebase app indexing plugin,
 version: 0.1.0
 homepage: https://github.com/wiseminds/firebase_app_indexing
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <3.5.0"
   flutter: ">=2.0.0"
 
 dependencies:
@@ -13,9 +13,9 @@ dependencies:
 
 dev_dependencies:
   flutter_test:
-    sdk: flutter 
+    sdk: flutter
 
-flutter: 
+flutter:
   plugin:
     platforms:
       android:
